calculation/src/requirements-fargate-compiled.txt
# WARNING: --strip-extras is becoming the default in version 8.0.0. To silence this warning, either use --strip-extras to opt into the new default or use --no-strip-extras to retain the existing behavior.
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
alabaster==1.0.0
    # via sphinx
babel==2.17.0
    # via sphinx
boto3==1.42.3
    # via
    #   -r requirements.in
    #   cmtools
botocore==1.42.3
    # via
    #   boto3
    #   s3transfer
certifi==2025.11.12
    # via requests
charset-normalizer==3.4.4
    # via requests
cmtools @ git+https://github.com/cloudmrhub/cloudmr-tools.git
    # via
    #   -r requirements.in
    #   mrotools
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
docutils==0.21.2
    # via
    #   sphinx
    #   sphinx-rtd-theme
fonttools==4.61.0
    # via matplotlib
h5py==3.15.1
    # via
    #   -r requirements.in
    #   ismrmrd
idna==3.11
    # via requests
imageio==2.37.2
    # via scikit-image
imagesize==1.4.1
    # via sphinx
ismrmrd==1.14.2
    # via raider-eros-montin
ismrmrd-python-tools @ git+https://github.com/ismrmrd/ismrmrd-python-tools.git
    # via raider-eros-montin
jinja2==3.1.6
    # via sphinx
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonpickle==4.1.1
    # via -r requirements.in
kiwisolver==1.4.9
    # via matplotlib
lazy-loader==0.4
    # via scikit-image
markupsafe==3.0.3
    # via jinja2
matplotlib==3.10.7
    # via
    #   cmtools
    #   mrotools
    #   pyable-eros-montin
    #   twixtools
    #   vtk
mrotools @ git+https://github.com/cloudmrhub/mroptimum-tools.git
    # via -r requirements.in
networkx==3.4.2
    # via scikit-image
numpy==2.2.6
    # via
    #   -r requirements.in
    #   contourpy
    #   h5py
    #   imageio
    #   ismrmrd
    #   matplotlib
    #   pyable-eros-montin
    #   raider-eros-montin
    #   scikit-image
    #   scipy
    #   tifffile
    #   twixtools
packaging==25.0
    # via
    #   lazy-loader
    #   matplotlib
    #   scikit-image
    #   sphinx
pillow==12.0.0
    # via
    #   cmtools
    #   imageio
    #   matplotlib
    #   mrotools
    #   scikit-image
pyable-eros-montin @ git+https://github.com/erosmontin/pyable.git
    # via
    #   cmtools
    #   mrotools
pygments==2.19.2
    # via sphinx
pygrappa==0.26.3
    # via cmtools
pynico-eros-montin @ git+https://github.com/erosmontin/pynico.git
    # via
    #   -r requirements.in
    #   cmtools
    #   mrotools
    #   pyable-eros-montin
pyparsing==3.2.5
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   matplotlib
raider-eros-montin @ git+https://github.com/erosmontin/raider.git
    # via
    #   cmtools
    #   mrotools
requests==2.32.5
    # via
    #   -r requirements.in
    #   cmtools
    #   sphinx
s3transfer==0.16.0
    # via boto3
scikit-image==0.25.2
    # via pyable-eros-montin
scipy==1.15.3
    # via
    #   cmtools
    #   mrotools
    #   scikit-image
    #   twixtools
simpleitk==2.5.3
    # via pyable-eros-montin
six==1.17.0
    # via python-dateutil
snowballstemmer==3.0.1
    # via sphinx
sphinx==8.1.3
    # via
    #   cmtools
    #   sphinx-rtd-theme
    #   sphinxcontrib-jquery
sphinx-rtd-theme==3.0.2
    # via cmtools
sphinxcontrib-applehelp==2.0.0
    # via sphinx
sphinxcontrib-devhelp==2.0.0
    # via sphinx
sphinxcontrib-htmlhelp==2.1.0
    # via sphinx
sphinxcontrib-jquery==4.1
    # via sphinx-rtd-theme
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==2.0.0
    # via sphinx
sphinxcontrib-serializinghtml==2.0.0
    # via sphinx
tifffile==2025.5.10
    # via scikit-image
tomli==2.3.0
    # via sphinx
tqdm==4.67.1
    # via twixtools
twixtools @ git+https://github.com/pehses/twixtools.git
    # via raider-eros-montin
typing-extensions==4.15.0
    # via xsdata
urllib3==2.5.0
    # via
    #   botocore
    #   requests
vtk==9.5.2
    # via pyable-eros-montin
xsdata==25.7
    # via ismrmrd

# The following packages are considered to be unsafe in a requirements file:
# setuptools